+++ packages/agent-app-sdk/src/widgets/details/details-widget.ts
@@ -1,4 +1,4 @@
-import { createConnection, createWidget, IConnection, withAmplitude, withPayments } from '@livechat/widget-core-sdk';
+import { createConnection, createWidget, IConnection, withAmplitude, withPayments, withToasts } from '@livechat/widget-core-sdk';
 import { withCustomerProfile } from '../shared/customer-profile';
 import { withRichMessages } from '../shared/rich-messages';
 import { withTheme } from '../shared/theme';
@@ -25,7 +25,7 @@ export function DetailsWidget(connection: IConnection<IDetailsWidgetEvents>) {
     }
   );
 
-  const widget = withAmplitude(withRichMessages(withCustomerProfile(withTheme(withPayments(base)))));
+  const widget = withAmplitude(withRichMessages(withCustomerProfile(withTheme(withToasts(withPayments(base))))));
 
   return widget;
 }