+++ packages/agent-app-sdk/src/widgets/fullscreen/fullscreen-widget.ts
@@ -1,4 +1,4 @@
-import { createConnection, createWidget, IConnection, withAmplitude, withPayments } from '@livechat/widget-core-sdk';
+import { createConnection, createWidget, IConnection, withAmplitude, withPayments, withToasts } from '@livechat/widget-core-sdk';
 import { withPageData } from '../shared/page-data';
 import { withTheme } from '../shared/theme';
 import { IFullscreenWidgetApi, IFullscreenWidgetEvents, ReportsFilters } from './interfaces';
@@ -29,7 +29,7 @@ export function FullscreenWidget(
     }
   );
 
-  return withAmplitude(withPageData(withTheme(withPayments(base))));
+  return withAmplitude(withPageData(withTheme(withToasts(withPayments(base)))));
 }
 
 export type IFullscreenWidget = ReturnType<typeof FullscreenWidget>;