+++ packages/widget-core-sdk/src/connection/driver.ts
@@ -9,12 +9,12 @@ import { getIsEventOriginAllowed } from './helpers';
 
 export function Plain(): IConnectionDriver {
   // eslint-disable-next-line @typescript-eslint/no-empty-function, @typescript-eslint/no-unused-vars
-  let currentListener: ConnectionListener = (_: IInboxMessage) => {};
+  let currentListener: ConnectionListener = (_: IInboxMessage) => { };
 
   function handleEvent(event) {
-    const isEventOrignAllowed = getIsEventOriginAllowed(event.origin);
+    const isEventOriginAllowed = getIsEventOriginAllowed(event.origin);
 
-    if (isEventOrignAllowed) {
+    if (isEventOriginAllowed) {
       currentListener(event.data);
     }
   }
@@ -39,7 +39,7 @@ interface ITrustedChild {
 
 export function Trusted(): Promise<IConnectionDriver> {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function
-  let currentListener: ConnectionListener = (_: IInboxMessage) => {};
+  let currentListener: ConnectionListener = (_: IInboxMessage) => { };
 
   function handleMessage(message: IInboxMessage) {
     currentListener(message);
@@ -52,7 +52,7 @@ export function Trusted(): Promise<IConnectionDriver> {
       },
       send(message: IOutboxMessage) {
         // eslint-disable-next-line @typescript-eslint/no-empty-function
-        return child.call('handle', message).then(() => {});
+        return child.call('handle', message).then(() => { });
       }
     })
   );