+++ packages/widget-core-sdk/src/toasts.ts
@@ -1,0 +1,24 @@
+import { WidgetMixin } from './widget';
+
+type ToastKind = 'success' | 'error';
+
+export interface IToast {
+  content: string;
+  kind: ToastKind;
+}
+
+export interface IWithToastApi {
+  setToast(): Promise<void>;
+}
+
+export const withToasts: WidgetMixin<IWithToastApi, void> = widget => {
+  return {
+    ...widget,
+    setToast(toast: IToast) {
+      return widget.sendMessage(
+        'set_toast',
+        toast
+      );
+    }
+  };
+};